refactor: move ssh-exec to servers dir
